refactor: bucketCount 컴포넌트에 있던 기능들 훅으로 분리
